requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
pika
requests
pytest
pytest-httpx